refactor(Expense Claim): extract child tables as separate components
