+++ frontend/src/components/ExpenseAdvancesTable.vue
@@ -1,0 +1,60 @@
+<template>
+	<div class="flex flex-row justify-between items-center">
+		<h2 class="text-lg font-semibold text-gray-800">Settle against Advances</h2>
+	</div>
+
+	<div class="flex flex-col gap-2.5">
+		<!-- Advance Card -->
+		<div
+			v-for="advance in expenseClaim.advances"
+			:key="advance.name"
+			class="flex flex-col bg-white border shadow-sm rounded-lg p-3.5 cursor-pointer"
+			:class="advance.selected ? 'border-blue-500' : ''"
+			@click="advance.selected = !advance.selected"
+		>
+			<div class="flex flex-row justify-between items-center">
+				<div class="flex flex-row items-start gap-3">
+					<Input type="checkbox" class="mt-0.5" v-model="advance.selected" />
+
+					<div class="flex flex-col items-start gap-1">
+						<div class="text-lg font-semibold text-gray-800">
+							{{ advance.purpose || advance.employee_advance }}
+						</div>
+						<div class="flex flex-row items-center gap-3 justify-between">
+							<div class="text-sm font-normal text-gray-500">
+								{{
+									`Unclaimed Amount: ${currency} ${advance.unclaimed_amount}`
+								}}
+							</div>
+						</div>
+					</div>
+				</div>
+
+				<div class="flex flex-row items-center gap-2">
+					<span class="text-normal">
+						{{ currency }}
+					</span>
+					<Input
+						type="number"
+						class="w-20"
+						v-model="advance.allocated_amount"
+						@input="(v) => (advance.selected = v)"
+					/>
+				</div>
+			</div>
+		</div>
+	</div>
+</template>
+
+<script setup>
+const props = defineProps({
+	expenseClaim: {
+		type: Object,
+		required: true,
+	},
+	currency: {
+		type: String,
+		required: true,
+	},
+})
+</script>
